+++ backend/src/main/java/com/handong/rebon/auth/presentation/LoginInterceptor.java
@@ -1,11 +1,13 @@
-package com.handong.rebon.auth.infrastructure;
+package com.handong.rebon.auth.presentation;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import com.handong.rebon.auth.domain.RequiredLogin;
+import com.handong.rebon.auth.infrastructure.JwtProvider;
 import com.handong.rebon.util.AuthorizationExtractor;
 
+import org.springframework.http.HttpMethod;
 import org.springframework.stereotype.Component;
 import org.springframework.web.method.HandlerMethod;
 import org.springframework.web.servlet.HandlerInterceptor;
@@ -20,11 +22,15 @@ public class LoginInterceptor implements HandlerInterceptor {
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
         HandlerMethod handlerMethod = (HandlerMethod) handler;
-        if (!handlerMethod.hasMethodAnnotation(RequiredLogin.class)) {
+        if (!handlerMethod.hasMethodAnnotation(RequiredLogin.class) || isPreflight(request)) {
             return true;
         }
         String token = AuthorizationExtractor.extractAccessToken(request);
         jwtProvider.validateToken(token);
         return true;
+    }
+
+    private boolean isPreflight(HttpServletRequest request) {
+        return HttpMethod.OPTIONS.matches(request.getMethod());
     }
 }
